backend/requirements.txt
fastapi
uvicorn
twilio
firebase-admin
pydantic
sqlalchemy
python-multipart
python-dotenv
httpx
